Se implementa cru usuarios
